[FEAT/#53] DialogContent 색상 버튼에 따른 UI 구현
