+++ core/src/main/res/drawable/ic_up_22.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="22dp"
+    android:height="22dp"
+    android:viewportWidth="22"
+    android:viewportHeight="22">
+  <path
+      android:pathData="M10.508,7.293C10.898,6.902 11.53,6.902 11.92,7.293L17.135,12.52C17.525,12.911 17.525,13.545 17.135,13.936C16.745,14.327 16.113,14.327 15.722,13.936L11.214,9.417L6.706,13.936C6.316,14.327 5.683,14.327 5.293,13.936C4.903,13.545 4.903,12.911 5.293,12.52L10.508,7.293Z"
+      android:fillColor="#ffffff"
+      android:fillType="evenOdd"/>
+</vector>
